+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Don't clean up stale storage data in railtie
+
 ## [0.5.0] - 2025-09-21
 
 - Add storage interface with Redis cluster and Memcached adapters for distributed deployments